docs: add documentation for modules, refactor public API
