+++ src/lib.rs
@@ -4,6 +4,18 @@
 //! 
 //! The library provides a simple API for making requests to websites, and it also allows you to customize the request headers, use proxies, custom timeouts and more.
 //! 
+//! ```rust
+//! use retch::retcher::Retcher;
+//! 
+//! #[tokio::main]
+//! async fn main() {
+//!    let retcher = Retcher::new().unwrap();
+//!    let response = retcher.get("https://example.com").send().await.unwrap();
+//! 
+//!    println!("{}", response.text().await.unwrap());
+//! }
+//! ```
+//! 
 //! ### Other projects
 //! 
 //! If you are looking for a command-line tool that allows you to make requests to websites, check out the [`retch-cli`](https://github.com/retch-http/retch-cli/) project.
@@ -13,14 +25,13 @@
 //! npm install retch-http
 //! ```
 //! 
-//! ### ⚠️ Warning ⚠️
+//! ### Usage from Rust
 //!
 //! Technically speaking, the `retch` project is a somewhat thin wrapper around `reqwest` that provides a more ergonomic API for making requests to websites. 
 //! The real strength of `retch` is that it uses patched versions of `rustls` and other libraries that allow it to make browser-like requests.
 //! 
 //! Note that if you want to use this library in your rust project, you have to add the following dependencies to your `Cargo.toml` file:
 //! ```toml
-//! 
 //! [dependencies]
 //! retch = { git="https://github.com/retch-http/retch.git", branch="master" }
 //! 
@@ -37,21 +48,29 @@
 #![deny(unused_crate_dependencies)]
 mod http_headers;
 mod tls;
-pub mod retcher;
-
-#[derive(PartialEq, Debug, Clone, Copy, Default)]
-pub enum Browser {
-  #[default]
-  Chrome,
-  Firefox,
-}
-
-pub(crate) mod http3;
-pub(crate) mod request;
 mod response_parsing;
 
-pub use request::RequestOptions;
+pub(crate) mod http3;
 
+pub mod retcher;
+pub mod request;
+
+/// Contains browser emulation-related types and functions.
+pub mod emulation {
+  
+  /// The `Browser` enum is used to specify the browser that should be impersonated.
+  /// 
+  /// It can be passed as a parameter to [`RetcherBuilder::with_browser`](crate::retcher::RetcherBuilder::with_browser) 
+  /// to use the browser emulation with the built [`Retcher`](crate::retcher::Retcher) instance.
+  #[derive(PartialEq, Debug, Clone, Copy, Default)]
+  pub enum Browser {
+    #[default]
+    Chrome,
+    Firefox,
+  }
+}
+
+/// Various utility functions and types.
 pub mod utils {
   pub use crate::response_parsing::decode;
   pub use crate::response_parsing::ContentType;
