+++ src/response_parsing/mod.rs
@@ -27,7 +27,7 @@ fn bom_sniffing(bytes: &Vec<u8>) -> Option<encoding::EncodingRef> {
 /// A lazy implementation of the BOM sniffing algorithm, using `scraper` to parse the HTML and extract the encoding.
 /// 
 /// See more details at https://html.spec.whatwg.org/#prescan-a-byte-stream-to-determine-its-encoding
-pub fn prescan_bytestream(bytes: &Vec<u8>) -> Option<encoding::EncodingRef> {
+fn prescan_bytestream(bytes: &Vec<u8>) -> Option<encoding::EncodingRef> {
     if bytes.len() < 4 {
         return None;
     }
@@ -59,7 +59,29 @@ pub fn prescan_bytestream(bytes: &Vec<u8>) -> Option<encoding::EncodingRef> {
     None
 }
 
-/// Converts a vector of bytes to a string using the provided encoding.
+/// Converts a vector of bytes to a [`String`] using the provided encoding.
+/// 
+/// If the encoding is not provided, the function tries to detect it using the BOM sniffing algorithm 
+/// and the byte stream prescanning algorithm.
+/// 
+/// ### Example
+/// 
+/// ```rust
+/// let bytes = vec![0x48, 0x65, 0x6C, 0x6C, 0x6F];
+/// let string = decode(&bytes, None);
+/// 
+/// assert_eq!(string, "Hello"); // By default, the function uses the UTF-8 encoding.
+/// 
+/// let bytes = vec![0xFE, 0xFF, 0x00, 0x48, 0x00, 0x65, 0x00, 0x6C, 0x00, 0x6C, 0x00, 0x6F];
+/// let string = decode(&bytes, None);
+/// 
+/// assert_eq!(string, "Hello"); // The function detects the UTF-16BE encoding using the BOM sniffing algorithm.
+/// 
+/// let bytes = vec![0x9e, 0x6c, 0x75, 0x9d, 0x6f, 0x75, 0xe8, 0x6b, 0xfd, 0x20, 0x6b, 0xf9, 0xf2];
+/// let string = decode(&bytes, Some(encoding::all::WINDOWS_1250));
+/// 
+/// assert_eq!(string, "žluťoučký kůň"); // The function uses the Windows-1250 encoding.
+/// ```
 pub fn decode(bytes: &Vec<u8>, encoding_prior_knowledge: Option<encoding::EncodingRef>) -> String {
     let mut encoding: encoding::EncodingRef = encoding::all::UTF_8;
 
@@ -74,6 +96,16 @@ pub fn decode(bytes: &Vec<u8>, encoding_prior_knowledge: Option<encoding::Encodi
     return encoding.decode(&bytes, encoding::DecoderTrap::Strict).unwrap();
 }
 
+/// A struct that represents the contents of the `Content-Type` header.
+/// 
+/// The struct is used to extract the charset from the `Content-Type` header and convert it to an [`encoding::EncodingRef`].
+/// 
+/// ### Example
+/// ```rust 
+/// let content_type = ContentType::from("text/html; charset=cp1250").unwrap();
+/// 
+/// decode(&bytes, content_type.into());
+/// ```
 pub struct ContentType {
     charset: String,
 }
